ASSAO can now also use relative include paths (to re-use DecodeNormalGBuffer)
